Release v1.2.0

+++ test/test-utils/clear.ts
@@ -1,6 +1,7 @@
 import { DataSource } from 'typeorm'
 
 const clearDb = async (dataSource: DataSource) => {
+  if (process.env.NODE_ENV === 'prod') return
   await dataSource.dropDatabase()
   await dataSource.synchronize()
 }
